Adds dependent-vector marking (dashed lines) in FOGI diagram.
FOGI quantities that are linearly dependent upon other FOGI quantities
that utilize the same number of operations can be retained and marked
instead of dropped in the construction of FOGI quantities.  The FOGI
diagrams have been updated to include such quantities using dashed lines.

This commit also fixes what I think amounts to a scaling bug in the
conversion between FOGI quantities and a model's error generator
coefficients.  Previously, we used the constructed matrix of FOGI
vectors to transition from FOGI coefficients -> errorgen coefficients,
but I think the columns of this matrix should be interpreted as
*directions* in error generator space (like their labels indicate),
and so really the transpose of the matrix transitions from errorgen
coefficients -> FOGI coefficients, and the psuedo-inverse of the
transposed matrix is what transitions the other way.