Adds better relational-FOGI-quantity metric and more FOGI plot options.
Introduces "jamiolkowski angle" as a potentially good metric for quantifying
relational Hamiltonian (coherent error) FOGI quantities, as it converts the
angle-like raw FOGI quantity into a average-gate-error fidelity-like "j angle"
number (defined via the diamond-norm-like action of a doubled map on an
entangled state but *without* the maximization).   A j-angle can be computed
for and of the gates that participate in the relational quantity, and we
currently just use the minimum of these the overall value (this is ad-hoc
and needs more thought).

This commit also expands the plotting options in the FOGIDiagram class,
adding methods to create heatmaps and a large latex table of detailed values.
This is all still experimental and should be cleaned up later - but probably
kept in a single class or base class as there's a lot of overlap between the
diagrams.