MASSIVE update that adds distributed memory foundation to GST calcs (WIP)
Previously multiple processors were utilized to compute circuit outcome
probabilities, but these where MPI-gathered up immediately afterwared,
so that each processor held the full number of computed probabilities
(elements) and parameters.  This commit begins to implement a new &
better paradigm whereby the memory for element, Jacobian, etc. arrays
is distributed among the processors.  This effectively removes all the
"gather" calls from within the probability calc routines and moves them
into separate layout/resource-alloc gather calls that the user can invoke
if they (ever) want the entirety of an array on the root processor.

The role of a DistributableCOPALayout is greatly expanded now, as it
encapsulates the mapping between "global" and "local" quantities.  This
is a WIP and there are numerous debug/REMOVE marked sections, as well
as some hacks to facilitate debugging, namely 1) the emulation of
multiple hosts when there is really just one (in ResoureAllocation)
and 2) an ad-hoc default tree splitting (# atoms) in
MatrixForwardSimulator.create_layout(...).  Furthermore, this commit
has only updated the matrix forward simulator - the "map" and "term"
variants still need to be updated (e.g. their calc routines need to
expect only *local* arrays and they don't need to gather anything).