Another large update that updates the custom-LM routine for distributed memory.
This commit adds an intermediate "distributed quantity calculator"
object, which may be best subsumed into the layout, that exposes an
API to the updated custom_leastsq routine so that it can work with
distributed memory.  Lots of minor bug fixes were applied during the
continued testing of the distributed-memory code, and now it seems
to work for 1Q GST tests on multiple processors.  Issues still to
iron out are:
- holding x as a distributed vector may be counterproductive
- the custom LM routine can't use its SVD-basis mode yet (TODO)
- map and term forwardsims still need updating
- still no Hessian support
- no "extra" elements (for penalty terms) support
- HACKs (on hostnames and layout creation) should be removed after testing
- lots of REMOVE-marked sections should be removed