Adds distributed-memory support for objfn penalties, Hessian, and SVD-basis.
A substantial update that fills out some of the distributed memory
functionality that was lacking up to this point.  Namely, this commit
adds support for computing the "extra elements" needed for objective
function penalties and Hessians (these were saved until last) using
a distributed-memory layout.  It also adds support for the SVD-defined
damping basis within the custom LM routine, though it's not very
computationally efficient (we gather the JTJ matrix and do a serial
eigensolve!).  Finally, this commit fixes a bug in distributedqtycalc.py
(a forgotten ".rank") that was causing mu to vary among the processors
during the custom LM optimization (and the opt to fail).  This commit
also re-introduces the shared-memory barriers needed in objective fn
methods, though now on firmer ground as there is a clear correspondence
between comms and common shared memory now (e.g. the "param-processing"
comm corresponds to the processors sharing the same segment of
the jacobian - ie an (element, param)-indexed - array).

The last remaining piece to the (initial) distributed-memory
implementation is the updating of the map and term fwd
simulators (currently on the matrix fwdsim can be used).