Simplifies layout creation; adds shared-mem sync. fixes.
Updates fwdsim classes to take new arguments that can control
the number of atoms, processor grid, and param-block sizes.  The
create layout functions now just take a single shot at getting a
good layout and fail if the memory limit is violated (but the mem
limit code is mostly broken now, so we need to fix this later on).

This commit also fixes shared-memory concorrency issues, adding
conditionals and barriers to ensure that only one the "lead" processor
writes to a section of shared mem AND the other procs wait for the
lead.  In ptic, the fwdsim.bulk_fill_* functions now don't return
until the proc is guaranteed to have its array filled, removing the
 need for this synchronization from objective fn methods like lsvec(...).

Also a bug in the distributed layout was fixed so that when a single
atom-processor has multiple atoms values are computed correctly.