Fixes support for memory estimation using distributed memory.
Updates the memory estimation framework to have local vs global
array counts distinct from the "atom" and "param_block" counts,
and updates the array types to indicate that most arrays are of
the local type.  This should fix the memory estimting code so that
it computes a roughly correct memory estimate (i.e. corresponding to
what is actually allocated) but it hasn't been tested yet.