Updates layout & fill_dprobs usage in fiducial pair reduction routines.
Some routines in fiducialpairreduction.py utilize layouts and models
to compute bulk probability-derivatives (dprobs).  These routines
explicitly don't use a comm (comm=None) so local arrays will be global
arrays, etc., but this commit updates the usage to make more explicit
that a local array should be constructed for bulk_fill_dprobs to act
on, and comments are added to remind us that these functions generally
work with local arrays that just happen to be global arrays in these
particular (comm=None) cases.