Fixes minor bug in param-slice ownership within a distributed layout.
We never actually encountered this, but the code for determining
ownership of a parameter slice (for assempling JTJ in fill_jtj(...))
could fail when there were lots of procs - namely when multiple procs
were assigned the same (element_slice, param_slice) part of the
Jacobian.  Now the root processor of this group alone is deemed the
"owner" of that parameter slice within the atom processor.