Removes virtual-host HACK in ResourceAllocation and adds env variable.
Now the environment variable PYGSTI_MAX_HOST_PROCS can be used to
add a "tag" the actual host name of a machine so as to create multiple
"virtual hosts" on only one actual host.  During debugging we found
this useful, and maybe we'll use it in testing in the future.