Updates the VolumetricBenchmarks.ipynb tutorial so it runs again.
This just involved some updates to the naming, as now the relevant
protocol is called `ByDepthSummaryStatistics`, and an update to the
pandas dataframe creation code so that pandas.Categorical dataframe
columns are created correctly when the data they hold is tuple-valued
(e.g. the column of which qubits are involved) -- we must explicitly
specify that these are object-type values otherwise the use of
numpy within pandas complains about ragged lists (issues a deprecation warning).