Relaxes an assert statement within the Wildcard newton solve.
The assertion checks that the Hessian of the objective function is
positive.  The update allows the hessian to have small negative
eigenvalues, where "small" is "at least 8 orders of magnitude
smaller than the largest eigenvalue".  This fixes an example case
where the assertion was (I think) falsely triggered by a -5e-4
eigenvalue when all others were ~ 1e20.