Updates "bad-fit" estimation so that a wildcard budget is always added, even if it's zero.
Previously and additional check in _compute_wildcard_budget (beyond a global check in
add_badfit_estimates whether a fit is indeed "bad") would short-circuit the wildcard
budget calculation based on the significance criteria and result in a Results object not having
a wildcard budget.  This is unexpected, and this commit "fixes" this issue so that even
when a budget is not needed based on the aggregate & per-circuit criteria a budget is returned
(with an all-zero wildcard vector).  In the future we may want to combine the badfit *threshold*
with the significance level used to determine when a budget is acceptable.