Adds new FOGI "grid" diagram for creating multi-detail-level error diagrams.
In our continued effort to depict the FOGI errors nicely, the new grid
diagrams do this well for 1-local and 2-local errors.  If higher-local
errors are rare these could be very useful.  3 levels of detail are
possible, showing different granularity of the errors.  Aggregated
errors between pairs of gates (or a single gate) can be decomposed
by either *weight* or *target qubits* (an argument to the new
plot_multiscale_grid function).