Fixes some minor wildcard bugs; adds wildcard_inadmissable_action badfit option.
Changes an exact == to a np.isclose call when checking whether frequencies are
zero or probabilities exactly equal frequencies within the wildcard budget computation.
Also adds a new `wildcard_inadmissable_action` attribute to GSTBadFitOptions, which
defaults to 'raise' (so an exception is raised when a budget is inadmissable) but
can also be 'print' (so that a warning is printed but analysis continues).