Updates/fixes ExplicitOpModel "reparameterize" option when setting up FOGI qtys.
Finishes up the work of creating an interposer object held by
an ExplicitOpModel (to convert between "operation" and "model" parameter
vectors) that causes the paramters of the model to become the FOGI
quantities themselves.  Currently there's no support to constrain these
quantities to being positive, and we've only tested it with base H+s
models so that there are no underlying positivity constraints on the
operation parameters - I think this is currently a requirement since
otherwise a derivative inverse won't exist when we need it to (within
the interposer construction code in ExplicitOpModel._add_reparameterization).