Speeds up custom solve and adds threshold for just using SciPy serially.
Factors the work done in custom_solve to facilitate profiling and
adds enhancements to speed up the worst bottlenecks so that the
custom_solve routine should be running about as fast as it can without
a major algorithmic overhaul.  An added serial_solve_proc_threshold
optimizer argument sets a #-of-processors threshold under which Scipy's
solve(...) routine is just called serially.

This commit also fixes a bug in the custom_solve routine whereby
the A and b values were modified upon exiting the function.  This
threw off the LM solve at least in some cases - now we see identical
behavior with custom_solve and using Scipy's solve, so it seems to
be correct now.