Updates and fixes use of prefix tables in MapForwardSimulator.
Previously, the `max_cache_size` element wasn't even plumbed up so
there was not way to actually set the value to anything other than
the default of `None`.  Still this ended up never creating a cache
of more than 0 elements because the circuits we initially sent to
PrefixTable.__init__ where *complete* circuits, i.e. then included
a terminating POVM.  This made every circuit a viable prefix for NO
others, and so even though we attempted to construct a prefix cache,
one was never actually utilized.

This commit remedies this by plumbing `max_cache_size` and adding
richer functionality to OpModel.split_circuit so that it just remove
the POVM off of a Circuit.  The POVM-less circuits are used to construct
a prefix table within the MapCOPALayout object, and after some additional
bug fixes (since the prefix-table splitting code never actually had to
run previously) PrefixTables can again be used by the map layout as intended.

This commit also adds performance enhancements to PrefixTable.__init__
so that when all evaluated circuits have the same line labels the
circuits are reduced to native tuple-of-circuit-layer-labels objects
that are much faster to slice and check for equivalence.