Propagates final MDC store to post-GST on-bad-fit actions.
Previously the post-GST work of performing "bad-fit" analysis
(e.g. wildcard budget compuation) would need to re-compute a layout
and objective function in order to evaluate the objective function
within Estimate.mistfit_sigma (i.e. when computing nSigma to see if
a fit was bad).  This is inefficient, and this commit utilizes the
saved final objective function object to compute nSigma when it's
possible to do so.

As a side effict, this commit updates the ObjectiveFunction.fn()
method to return the *global* (summed across processors) objective
function value and creates a new fn_local() method that returns
the local contribution (though it's unlikely this will be needed).

This also required updating the customLM method so it evaluates
the objective function one final time at its end if necessary
to ensure that the last objective function call was made at the
optimal x (param-vector) value.