Updates the default number of FD iterations to 0 for map forward sims.
Previously the default number of finite-different (FD) optimizer
iterations was 1 except in the case of a term-fwdsim.  Making an initial
FD iteration is useful in the matrix-fwdsim case as the analytic
derivative at a singular starting (e.g. target) model could cause an
unhelpful first step, but the map-fwdsim already computes derivatives
using the FD method (but in a more efficient way), so there is no need
for it to set the "num_fd_iters=1" by default.  This commit changes the
default so it's 1 only for the matrix-fwdsim and 0 otherwise.