Updates shared memory structure to attempt to eliminate contention.
It seems that shared memory blocks can only be written to by one
processor at a time (even if different areas of the block are
being written to), so to increase performance this commit changes
the way shared memory blocks are allocated by a DistributableCOPALayout.
Previously, a single per-host (large) array was allocated and views
into this array were held by each processor (note: multiple procs
can hold the same view when they compute the same region of the array).
Now separate shared blocks are allocated, and a dict of the other
blocks is held so that the .host_array attribute of a LocalNumpyArray
can be accessed similarly as a single large array.  Since slices
cannot be hashed in Python, we use the new slicetools.slice_hash
function to convert a slice into a hashable (start, stop, step)
tuple.

This commit also removes some of the debugging structures of some
recent commits, so that now DM_mapfill_dprobs_block assigns directly
to it's array_to_fill argument (which will be a shared array when
shared memory is used).  The hope is that this is fine, and the writes
will happen quickly now because that shared memory block will only
be being written to by one processor (the "shared memory leader").