Changes hasattr(model, "num_nongauge_params") to test in dir(model) for performance.
Because num_nongauge_params is a method of a model, checking for its
existence with hasattr is inefficient because it actually runs the
property method when the property exists.  Since computing the number
of gauge/nongauge params can be a somewhat time consuming task, we don't
want to do it twice! (once in the hasattr check and once in the actual
function body of num_modeltest_params)