Adds 'return_shared' argument to layout's gather_local_array function & downstream.
Allows the caller of a COPALayout's gather_local_array (& similar functions)
to specify return_shared=True and have the method return the gathered array
as (potentially) a shared-memory array along with its handle.  This avoids
a numpy.copy operation at the end of the function, which can give a little
speed up to the gather_jtj and gather_jtf functions used before calling
scipy.linalg.solve in customsolve.py (when we just gather & call SciPy).