Updates and clarifies normalizations on FOGI quantities.
FOGI direction vectors are now normalized so their corresponding vector
(just the direction / L2norm(direction)**2 ) has a desired normalization.  In
the Hamiltonian error case this is just that the 2-norm == 1, so the direction
and vector are the *same* and things are nice and simple.  In the "other"
error case we set the L1-norm == 1, which seems right for stochastic errors
but probably not when we extend this to CPTP errors, etc. (FUTURE!).  Comments
have been updated to explain these normalizations, and that we don't necessarily
construct an orthogonal (or even independent when action == "mark") set of
FOGI direction vectors.  We now refer to the quantities dot(fogi_dir.T, errorvec)
as "components" rather than "coefficients", though the components are labelled
by the corresponding FOGI vector as if the direction was a part of an orthogonal
basis (so the vector, as labelled, is always just a scaled version of the direction).

Note that when moving from a set of fogi components to an errorgen vec, i.e. in
set_fogi_components, the pseudoinverse is used - and so the actual action will
only be consistent with the labelling if the fogi vectors are orthogonal.

When components are added to create summary numbers in a FOGI diagram, we
again assume that that fogi vectors being combined are orthogonal and add
the corresponding error *vectors* even if this isn't true.