Fixes accidental variable clobbering in ForwardSimulators bulk_probs(...)
Previously we had a separate shared memory handle to free the shared memory
of a LocalNumpyArray so we could clobber (reuse) the "vp" variable in
bulk_probs(...) with impunity.  We changed that recently, so this commit
avoids the variable name reuse by renaming vp => local_vp and similarly
for bulk_dprobs and bulk_hprobs functions.