Plumbs the sparse_lindblad_reps argument to a CloudNoiseModel more consistently.
Previously, this argument would control whether LindbladOps had dense representations
in *some* places, but would be silently overidden or replaced by a test of
whether the forward simulator was a MatrixForwardSimulator instance.  While this
behavior may have been convenient, it's much less transparent to the user.  Now
sparse_lindblad_reps is always used to determine the representation type, and
assertions guard against using sparse reps in conjunction with the matrix fwdsim
(disallowed because the latter can't handle them, or maybe it's just very
 inefficient? - we should look into this more).