Cleans up unnecessary 'resource_alloc' arguments.
Now that the layout has taken ownership of its own ResourceAlloc
(distributed layouts rely on a particular resource alloc), this commit
adds resource_alloc() as a method for *all* COPA layouts and in turn
removes resource_alloc as an argument alongside 'layout' to fwdsim
methods as it's now redundant.

This commit also neatens up the methods of DistributableForwardSimulator
so that the methods that specifically work on atoms end in _atom rather
than re-using (and clobbering) the _block function of ForwardSimulator.
This makes it clearer what is being done, and leaves room in the future
to implement the used (but not implemented yet!) .as_layout(resource_alloc)
method of a DistributableAtom.  Our currently used fwdsims (Matrix, Map,
and Term) override the now "_bulk_fill_?probs_atom" functions of their
base class, taking it upon themselves to efficiently extract what
circuits need to be computed and at what element positions the results
go.  The base class (DistributableForwardSimulator) now implements these
*_atom functions so that if the layout's atoms can be converted (even
if it means they're less efficient) to fully-fledged COPA Layouts then
a (new) forward simulator class would only need to implement the
circuit computation routine(s) and still be able to make use of a
distributed layout (currently a new fwdsim would need to contruct it's
own distributed layout type or be left with just the simple serial layouts).
 We leave this to future work.