First pass at stubbing out classes for integration of CHP as a ForwardSimulator.
Tim's stdgate_to_chp was reworked to fit in tools.internalgates.
A basic CHPOp added to objects.operation, along with CHPOpRep,
which should be reworked and moved to replib later
(should probably just mirror the QProg struct in chp.c).

A new WeakForwardSimulator class was added to serve as a base class
for the new CHPWeakForwardSimulator,
with the plan to wrap this into a standard ForwardSimulator later.