Fixes bug in wildcard budget probability updating.
A boundary case in which sets A, B, C, and D are *not* disjoint
caused this issue.  The problem was that we were making a strict
(zero tolerance) comparison for sets A and B but set C was using
np.isclose with its default tolerance, allowing elements to be both
in set C and in set B (or A).

This commit also updates the assertion message so that it's easier
to debug these boundary cases post-mortem.