Relaxes an assertion statement checking the allocation of shared memory.
It seems that shared memory has a minimum allocation size, which means that
when a very small abount (< 4096 bytes in our test case) of shared memory is
allocated the .size of the shared memory will be *greater* than the amount
requested.  Previously we asserted that these must be equal (to catch
out-of-sync processors allocated shared mem), but this behavior means we
need to relax the assertion to a >= test.