Adds a fail-safe to the prefix-table splitting heuristic.
It's possible, as we've recently observed, for the heuristic used to
intelligently split a prefix table into many subtables to fail.  This
commit adds a fail-safe so that after 100 iterations of trying to find
a maximum cost & rate for the table splitting, instead of assert-failing
we print a warning message and force the creation of the correct number
of sub-tables.