Updates, fixes, and clarifies distributed Hessian calculation.
Renames _singleatom and hprobs_by_blocks methods within the distributable
fwdsim to more accurately named iter_* methods (when a generator is returned)
and bulk_fill_*_atom methods (when one or more passed arrays that are already
sized for an atom need to be filled).  Names within the unit tests have
been update but not actually run.  Also, the iter* methods end in _by_rectangle
rather than "by_block", since we use block generally to mean a slice along a
single axis (a rectangle is a (block1, block2) tuple).  Clarifying the names
and interface fixes a bug in which, during the Hessian computation, an
array was indexed by an atom element slice twice, resulting in a shape
mismatch error.