Enables and fixes a bug in MPI tests.
The previous commit fixed some of the failures in the MPI tests,
and this commit un-comments a couple calls that were failing (computing
the Hessian using a multi-atom matrix sim and using param_blk_sizes).
It also updates the way we test parallel vs. serial computations by
comparing with the serial layout's .global_layout.  This seemed
unnecessary to me initially, but when there are multiple atoms the
"local" layout, even when there's just a single processor, is allowed
to shuffle the circuits around to facilitate computations.  This means
that the serial-local-layout may not have the circuits in the order of
the original circuit list we use to drive the comparison.  Its .global_layout
does.  A comment has been added to this effect in testMPI.py.