Updates custom linear solver to guard agains Scipy linsolve exceptions.
When the custom_solve routines falls back to call SciPy on just the
root proc, it needs to check for a LinAlgError exception and keeps
all the processors in sync.  This commit adds an additional broadcast
that communicates the success/failure of the scipy.linalg.solve to the
rest of the processors so that they all raise LinAlg error together.