Fixes a minor bug in PrefixTable splitting.
The loop used to construct a set of candidate tables was unnecessarily
initialized to a sub-table containing the first table elmement (circuit).
This was incorrect - it should have been initialized to an empty set,
and this can cause an assertion fail when trying to construct the
same number of tables as their are circuits.