Adds minor improvement to "tuned" method within DataSet.degrees_of_freedom.
This means of computing the actual degrees of freedom still seems a
little off, as a 10Q test's datagen model has a 2DeltaLogL corresponding
to N_sigma ~= 6 which seems to indicate the "tuned" degrees of freedom
are still underestimating the actual number.

See Erik's 2020 notebook "Efficient-Tomography-LDRD-SAND-figs.ipynb"
for derivation details.