Adds MAX_EXPONENT to prevent (some) exp overflows from occurring.
When running many (10 in our test case) qubit model fits a bad
optimizer step can result in sum-of-path-magnitudes and their max
(the exp(magnitude_sum)) to overflow.  This commit adds a guard that
limits the max-sum-of-path-magnitudes to be near (actually 10 orders
of magnitude less) than the maximum allowed floating point value.
There are still issues with overflows within the replibs, I think
when computing the actual sum-of-path-magnitudes (SOPM) in these
cases -- a FUTURE todo item.