Fixes more edge-case behavior of wildcard budget probability updates.
In certain cases processing the p/f ratios in their original sorted
order is incorrect - this is because the (alpha, beta) tuple moves
along a sort of path away from (1.0, 1.0) to move fvec toward qvec,
but alpha and beta are related to one another as a function of the
total (sum) of the frequencies in the A or B "sets" and this means
that a given element (of fvec and qvec) must have it's alpha *and*
beta considered and the closest alpha or beta to 1.0 should be chosen.

The fix is implemented in a fairly brute force way, finding the best
element on each iteration - so this commit makes the wildcard budget
calculations slower.  This commit also leaves lots of debugging and
checking code that slows things further.