Improves wildcard budget update_probs for performance.
Gives a 2-3x speedup to update_probs routine, which is the bottleneck
in wildcard error computations.  This commit updates the algorithm in
update_probs (though the clearer and slower update_circuit_probs
function) to be more efficient, and adds another pre-computed-info
argument to speed up repeated calls to update_probs that use the
same input probabilities and frequencies.  This additional precoputation
is now utilized by the barrier method but not by any other wildcard
method (yet).  Doing so should be pretty straightforward.

To further increase performance, we should write a cython version
of the update_probs method, as profiling indicates that most of
the time is spent inside the update_probs function.  The updated
algorithm is more amenable to this, but we leave a Cython version
as FUTURE work.