Adds a little more bookeeping to the FOGI-store object.
By storing the gauge-space gauge directions we can use the store
to construct the gauge operator that would (in the FOGI picture)
eliminate all non-FOGI (or "FOGV") component of an errorgen-set
vector.

Also adds DUAL NORM in several places to mark where we normalize a
"direction"/dual vector by taking the norm we want for the primal
vector and dividing it by the vector's norm**2.