Adds minor update to FOGI qty aggregation in FOGI diagram.
I think we were missing a normalization factor in the code that
aggregates the Hamiltonian quantities for diagram display.  This
wasn't an issue since the Hamiltonian case uses the L2 norm to normalize
its gauge-space vectors anyway - this was just uncovered by working
through the math and noticing a discrepancy with the code.  The
updated code is also more general, and handles the (non-existent so
far, I think) case when the gauge directions are not orthogonal.