Combines vis-network.js and raw SVG versions of the FOGI graph diagram.
Creates new classes: FOGIVisNetworkGraphDiagram and FOGISvgGraphDiagram which
inherit from base class FOGIGraphDiagram.  This update consolidates common
code between the two techniques, which will be helpful for future code
maintenance.

This commit also adds an ability to render numerical labels on the nodes
of a graph diagram but not the edges.