Upgrades time-dependent calcuations to use distributed memory paradigm.
Previously the updates to the distributed layout and calculations broke
the time-dependent calculations.  This fixes that.  Now, time-dependent
map-based calcs work as they did before (and should generally, but they
can be slow).  Matrix-based calcs no longer use multiple atoms for
different timestamps; instead, a single atom, atom-processor, and
param-processor are setup in create_layout, and then in the _timedep_
methods within matrixforwardsim.py the relevant atom-processing or
param-processing comm (the latter in the case of the jacobian) is
split to handle the timestamps in parallel.  Because there's only one
atom & param group this means all the processors are split to handle
different time steps, but in theory this should would even if we
change create_layout to allow for multiple atoms, etc. - but this
hasn't been completely thought through and there may be some issues,
especially with shared memory usage, that need to be worked out.

In particular, the timestamp_processing_ralloc comm should be all
the processors with the same set of (local) atoms and possibly
parameter range (in the case of Jacobian computation), so this
seems right, but we loop over the timestamps first then over the
atoms in the _timedep_ methods, so we should double check this.