Fixes a minor bug in wildcard dealing with handling of sum(probs) != 1 case.
In the WildcardBudget.update_probs method a check is performed that
adjusts the "qvec" (the vector of circuit outcome probabilities) so that
it's non-negative and now, after this commit, so that it sums to 1.0.
Previously if we adjusted qvec we in error didn't recompute all of the
precomputed values like the sum_XX values and `iA` and `iB` arrays that
depended on qvec.  This commit corrects this bug, and adds the same
qvec-adjustment logic to the standalone update_circuit_probs function.