Adds the `DataSimulator` class and `StateFunctionDataSimulator` class.
An initial implementation of a new type of object within the "protocols"
framework: a *data simulator* is run on an experiment design and produces
a ProtocolData object.  The new StateFunctionDataSimulator can use
custom functions to generate auxiliary information based on the simulated
*states* (not just the probabilities) to be stored in the
dataset (within the ProtocolData object) of simulated data.  This
makes it easy to use a model to generate simulated data and compute simple
meta-data to tag/identify/evaluate these circuits.

This commit also adds the `auxinfo_dataframe` method to `ProtocolData`
so that it's easy to access/dump the generated auxiliary information.