Adds methods of promoting experiment designs and assigning to a CombinedExperimentDesign.
New `promote_to_combined` and `promote_to_simultaneous` methods of an experiment
design create a combined or simultaneous design with the original edesign as its
single sub-design.  Also adds a the ability to assign new sub-designs to a
CombinedExperimentDesign so long as the added sub-design doesn't add any new
circuits.