Adds local optimization to improve residual-OVD computed at residual-TVD point.
We settled on computing the residual-OVD by evaluating it at the best
"T" (transition) matrix for the residual-TVD (since finding the latter
is a convex program).  This additional local optimization tries to
improve the T-matrix locally to achieve a better (lower) OVD.  A
warning is printed if this local optimization improves the OVD too
much, since this may indicate that the initial point was bad and
the final result isn't a global optimum.