Improves performance of data loading.
Circuit parsing for large hierarchies of edesigns/data/results
objects was targeted in particular.  Additions include:
- global caching for circuit parsing, so that parsing isn't repeated
  when unnecessary.
- New LabelTupWithTime and LabelTupTupWithTime classes allow the
  non-time version to be created much more quickly as these time-less
  label types can just use tuple's __new__ directly.
- Ability to give a preloaded_data argument to ProtocolResultsDir.from_dir
  can speed up loading of results when the data has already been loaded.
- updates to parse_datafile which BREAK backward compatiblity (in that
  frequency and "total count" type columns are no longer supported)
  increase its performance by accessing new low-level "add" functions
  of the DataSet object.

The main bottleneck is still in the parsing of circuits, but this
code is already written in Cython and probably can't be sped up too
much more.  The next slowest piece is parsing the data - the parse_dataline
funtion of stdinput.py to be precise - and perhaps some minor gains
could be had by moving this function to Cython in the FUTURE.

Also note that the corresponding multi-dataset loading code has
not been updated yet, and that's something we likely want to do
in the future.