Adds "freeform" edesign and data simulator.
Introduces an improved, and maybe final, API for computing per-circuit
quantities without needing a data set.  The meaning of the term "data set"
is now expanded to include arbitrary quantities attached to circuits, and
the new FreeformDataSet holds exactly that - an arbitrary collection of
per-circuit information.  Data-simulators naturally should be the
objects to construct freeform data sets, and so the data simulator
objects are able to compute custom per-circuit quantities (not just
outcome counts).  The ModelFreeformSimulator makes it easy to derive
data simulator classes that compute per-circuit quantities based on
the circuit simulations of model(s).  The FreeformDesign holds a
list of arbitrary circuits along meta/aux data for them.  Many of
these objects can be converted to/from pandas dataframes to facilitate
workflows that do lots of processing that doesn't need to carry
Circuit objects along with it (where string reps are fine).