Makes a minor update to StaticStandardOp unit tests.
Unit tests ran through all standard gate unitaries, assuming them all
to be matrices.  Recently we added to this list, however, a *function*
that generates a unitary (for an arbitrary-angle z-rotation gate), which
breaks the tests.  For now, we just ignore this case, and test only the
single-valued standard gates.