Updates a memory-limit-testing system test.
A test was failing due to settign a memory limit based on the current
memory usage, and now this should use a static limit.  The purpose of
this test has diminished since now we do not adjust the number of atoms
(sub-trees) based on the memory limit, though we may revive this functionality
somewhat in the future.