Updates Residual-OVD to just be a Residual-TVD scaled by OVD/TVD.
After some further testing, we have realized that the OVD doesn't
seem compatible with the foliation we used to move from TVD -> ResidualTVD.
In particular we find that the residual-OVD values, when computed
similarly to residual-TVD, can be weirdly behaved as a function of
spam error.  This is at least in part due to the fact that the "T"
(transfer mx) found by the residual-TVD computation may exactly match
the TP and Q distributions on the 0000 and 1111 bit strings, leaving
the only remaining discrepancies on bit strings without support in P0.

To fix this issue, this commit redefines the residual-OVD to just be
a scaled version of the residual-TVD (scaled by the total OVD/TVD ratio,
as the OVD is still a robust metric).