Adds `parameter_labels` attribute to Model and ModelMember objects.
This functionality provides a way to associate a human-readable label
with each of a model's or operator's parameters, accessible via the
`.parameter_labels` property.  Currently, a Model automatically generates
parameter labels by creating (operator_label, description_str) tuples
based on the description strings coded into the individual operators.
A Model's parameter labels may be updated/changed via the set_parameter_label
method, but an operator's labels currently cannot be (it's not clear
this is useful).

This functionality should be useful for manipulating the way models
are parameterized as well as sampling random noise models.