Adds a warning and safeguard against getting NaN for the 2DeltaLogL wildcard threshold.
When the number of model parameters is greater than the number of dataset
degrees of freedom the wildcard calculation code would get a NaN when
computing the threshold 2DeltaLogL value (because k was negative).  This
commit lower-bounds k at 1 and prints a warning when k would have been 0 or negative.