Updates __repr__ of Label objects so it evaluates back to the Label.
Previously a Label's repr() would give something like Label[('Gx',0)].
While the use of square brackets helped to distinguish the parentheses
in the label data, it meant that the repr() couldn't be simply cut
and pasted to generate the label (assuming Label was imported).  This
commit "fixes" that, and updates the Label __repr__ methods so that
the returned string can be used to construct a new label, e.g.
"Label(('Gx',0))".  A unit test is added to ensure this is the case.