Updates system tests that were broken after recent feature merges.
A number of them broke because we don't support "frequency" data
columns anymore (this increases the read-in performance).  Some others
broke because the label objects changed somewhat and so reading in
old pickle files didn't work - to fix this, this commit regenerates
the "reference" datasets & models used in the system tests.