Moves term-calc code back to op classes; updates LindbladErrorgen class.
Methods that compute terms are heavily intertwined with an operation's
parameterization, since the terms can contain polynomials of the operation's
parameters.  It seems best, therefore, to keep these methods in the operation
classes and use a single "term" evotype to hold cached info and the sub-evotype
used when constructing the term representations themselves.

This commit also performs a first bass on the LindbladErrorgen class,
moving evotype specific code into the evotype files.